+++ edisgo/flex_opt/reinforce_grid.py
@@ -180,17 +180,17 @@ def reinforce_grid(
     logger.debug("==> Check station load.")
 
     overloaded_mv_station = (
-        pd.DataFrame() if mode == "lv" else checks.hv_mv_station_load(edisgo_reinforce)
+        pd.DataFrame(dtype=float) if mode == "lv" else checks.hv_mv_station_load(edisgo_reinforce)
     )
 
     overloaded_lv_stations = (
-        pd.DataFrame() if mode == "mv" else checks.mv_lv_station_load(edisgo_reinforce)
+        pd.DataFrame(dtype=float) if mode == "mv" else checks.mv_lv_station_load(edisgo_reinforce)
     )
 
     logger.debug("==> Check line load.")
 
     crit_lines = (
-        pd.DataFrame() if mode == "lv" else checks.mv_line_load(edisgo_reinforce)
+        pd.DataFrame(dtype=float) if mode == "lv" else checks.mv_line_load(edisgo_reinforce)
     )
 
     if not mode or mode == "lv":
@@ -245,7 +245,7 @@ def reinforce_grid(
 
         logger.debug("==> Recheck station load.")
         overloaded_mv_station = (
-            pd.DataFrame()
+            pd.DataFrame(dtype=float)
             if mode == "lv"
             else checks.hv_mv_station_load(edisgo_reinforce)
         )
@@ -256,7 +256,7 @@ def reinforce_grid(
         logger.debug("==> Recheck line load.")
 
         crit_lines = (
-            pd.DataFrame() if mode == "lv" else checks.mv_line_load(edisgo_reinforce)
+            pd.DataFrame(dtype=float) if mode == "lv" else checks.mv_line_load(edisgo_reinforce)
         )
 
         if not mode or mode == "lv":
@@ -464,7 +464,7 @@ def reinforce_grid(
     logger.debug("==> Recheck station load.")
 
     overloaded_mv_station = (
-        pd.DataFrame() if mode == "lv" else checks.hv_mv_station_load(edisgo_reinforce)
+        pd.DataFrame(dtype=float) if mode == "lv" else checks.hv_mv_station_load(edisgo_reinforce)
     )
 
     if mode != "mv":
@@ -473,7 +473,7 @@ def reinforce_grid(
     logger.debug("==> Recheck line load.")
 
     crit_lines = (
-        pd.DataFrame() if mode == "lv" else checks.mv_line_load(edisgo_reinforce)
+        pd.DataFrame(dtype=float) if mode == "lv" else checks.mv_line_load(edisgo_reinforce)
     )
 
     if not mode or mode == "lv":
@@ -528,7 +528,7 @@ def reinforce_grid(
 
         logger.debug("==> Recheck station load.")
         overloaded_mv_station = (
-            pd.DataFrame()
+            pd.DataFrame(dtype=float)
             if mode == "lv"
             else checks.hv_mv_station_load(edisgo_reinforce)
         )
@@ -539,7 +539,7 @@ def reinforce_grid(
         logger.debug("==> Recheck line load.")
 
         crit_lines = (
-            pd.DataFrame() if mode == "lv" else checks.mv_line_load(edisgo_reinforce)
+            pd.DataFrame(dtype=float) if mode == "lv" else checks.mv_line_load(edisgo_reinforce)
         )
 
         if not mode or mode == "lv":
